services/restapi/requirements.txt
Django==2.2.5
django-cors-headers==3.1.0
djangorestframework==3.10.3
dbus-python==1.2.12
git+https://github.com/emlid/systemd-manager.git